requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
django-cors-headers==3.5.0  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django==3.1.1             # via -r requirements.in, django-cors-headers, django-filter, djangorestframework, djangorestframework-simplejwt
djangorestframework-camel-case==1.2.0  # via -r requirements.in
djangorestframework-simplejwt==4.4.0  # via -r requirements.in
djangorestframework==3.12.0  # via -r requirements.in, djangorestframework-simplejwt
idna==2.10                # via requests
pyjwt==1.7.1              # via djangorestframework-simplejwt
pytz==2020.1              # via django
requests==2.24.0          # via -r requirements.in
sqlparse==0.3.1           # via django
urllib3==1.25.10          # via requests
